several nodes joining and leaving

+++ tcp_functions.h
@@ -5,8 +5,8 @@ int tcp_server(char *ip, int port);
 
 int tcp_client (char *ip, int port);
 
-ssize_t send_tcp_message(int fd, char*message, int message_size);
+int send_tcp_message(int fd, char*message, int message_size);
 
-ssize_t receive_tcp_message(int fd, char* buffer, int buffer_size);
+int receive_tcp_message(int fd, char* buffer, int buffer_size);
 
 #endif
